Redo video cleanup
